fix cast warning

+++ src/windows/maze_construction.c
@@ -190,7 +190,7 @@ static void window_maze_construction_entrance_mouseup(rct_window *w, int widgetI
 		return;
 
 	RCT2_GLOBAL(0x00F44191, uint8) = widgetIndex == WIDX_MAZE_ENTRANCE ? 0 : 1;
-	RCT2_GLOBAL(0x00F44192, uint8) = w->number;
+	RCT2_GLOBAL(0x00F44192, uint8) = (uint8)w->number;
 	RCT2_GLOBAL(0x00F44193, uint8) = 0;
 	RCT2_GLOBAL(RCT2_ADDRESS_INPUT_FLAGS, uint32) |= INPUT_FLAG_6;
